add: jpa repository
- 테스트 jpa로 변경
- notification, memberNotification 저장, 조회 테스트 완료
- 테스트 fail 해결
