+++ .gitignore
@@ -36,3 +36,4 @@ out/
 
 ### VS Code ###
 .vscode/
+src/main/resources/data.sql
