+++ src/main/java/com/lotdiz/notificationservice/repository/MemberNotificationRepository.java
@@ -1,0 +1,9 @@
+package com.lotdiz.notificationservice.repository;
+
+import com.lotdiz.notificationservice.entity.MemberNotification;
+import com.lotdiz.notificationservice.entity.id.MemberNotificationId;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberNotificationRepository extends JpaRepository<MemberNotification, MemberNotificationId> {
+
+}
